Add new edge extraction filter
The new filter is more accurately approximates an angle between
normals to the interface when calculating surface functions.
An old filter is left for images with not so good resolution.

This change breaks API a bit. `edgemode` parameter to surface
functions is replaced with `filter` parameter of type `EdgeFilter`.
`EdgeFilter` can be instantiated like so:
`EdgeFilter(edge_5x5, BCPeriodic)`.

Filtering using GPU is disabled again and will be re-enabled in
future commits.