Fix F_{sss} function.
The first implementation was made to satisfy equations (11) in the
paper "Higher-order correlation functions in disordered media:
Computational algorithms and application to two-phase heterogeneous
materials" by Hessam Malmir, Muhammad Sahimi and Yang Jiao.

Turns out the paper is SEHR HÜBSCH, but neither (11) nor (12) is
true because F_{sss} has a unit of measurement m^-3 and F_{ss} has
a unit of measurement m^-2. They cannot be equal in any sence! It
took me a while to understand it. The same is true for F_{ssv}
(m^-2) and F_{sv} (m^-1).

Moreover the authors do not provide any way to verify correctness of
their algorithm with exception of the known analytic solution for
overlapping spheres. But again, this solution (which almost every
author uses in every paper, this is like 1 + 1 = 2 in the branch of
porous media) is for F_{ss} function and not for F_{sss}.

So instead of this solution (which is unapplicable to F_{sss} anyway)
I test F_{sss} calculated for a rotated cube. The analytical solution
is very easy to get using S. Torquato's method described in his book
"Random Heterogeneous Materials". This is not a very good test, but
it's OK for starters.